+++ pkg/util/k8sutil/resources/selectors.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2024 ArangoDB GmbH, Cologne, Germany
+// Copyright 2024-2025 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -21,9 +21,9 @@
 package resources
 
 import (
-	meta "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"slices"
 
-	"github.com/arangodb/kube-arangodb/pkg/util/strings"
+	meta "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 func SelectLabels(selector *meta.LabelSelector, labels map[string]string) bool {
@@ -46,7 +46,7 @@ func SelectLabels(selector *meta.LabelSelector, labels map[string]string) bool {
 
 			if v, ok := labels[req.Key]; !ok {
 				return false
-			} else if !strings.ListContains(req.Values, v) {
+			} else if !slices.Contains(req.Values, v) {
 				return false
 			}
 		case meta.LabelSelectorOpNotIn:
@@ -55,7 +55,7 @@ func SelectLabels(selector *meta.LabelSelector, labels map[string]string) bool {
 			}
 
 			if v, ok := labels[req.Key]; ok {
-				if strings.ListContains(req.Values, v) {
+				if slices.Contains(req.Values, v) {
 					return false
 				}
 			}