+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - (Feature) (Platform) Chart Overrides
 - (Feature) Parallel Executor
 - (Feature) (Platform) Service Handler
+- (Feature) (Platform) Service Handler Implementation
 
 ## [1.2.49](https://github.com/arangodb/kube-arangodb/tree/1.2.49) (2025-06-17)
 - (Maintenance) Optimize go.mod